Initial Commit
Initial upload of source code for 2014 JPL UIUC Space Design
Competition. The objective was to parse images and 3D height data to
find safe landing sites for a spacecraft. This implementation utilized
JAVA, JAMA, and JME to produce a geometric simulation based solution
which purely focuses on 3D height data.